Add some checksum code
One question I wanted to verify was that we were correctly capturing the `Body`
out of logs. One way to verify this is to check the checksum - that's what it's
there for, after all.

Rather than running the log verifier program though, I can actually just
directly compute the checksum in Clickhouse. There's a CRC32 function, but one
gap was the varint encoding used in Boulder to output the hash. It's annoying
but possible to recreate the function. I had the Junie agent to do it for me.

This isn't really useful but I'm putting it here, in the poc-poc repo, for
reference. Realistically, if we want to go this path, I'd rather change Boulder
to emit a checksum that's easier to validate.